OBI-Warp: use standard vector as data containers
OBI-Warp seems to have been written with efficiency as an important
factor. This is also probably the reason behind the authors using
manual data allocation and memory management, even though C++ (and
its STL) provides the utilities to ease some of this heavylifting.

El-MAVEN has been encountering multiple instances of deallocation
of invalid (unallocated or already deallocated) raw array pointers,
which leads to segfault crashes. This effect seems to be especially
more pronounced on Windows systems.

This patch changes much of the underlying data structures to use
safer vector templates over raw data arrays. There has been a
noticeable decline in performance of the algorithm runtime, but
the aforementioned crashes should no longer be an issue.